conferenti-ai-agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
azure-ai-agents==1.1.0
azure-common==1.1.28
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.1
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-monitor-opentelemetry==1.3.0
azure-monitor-opentelemetry-exporter==1.0.0b44
azure-search-documents==11.6.0
build==1.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
fastapi==0.109.2
fixedint==0.1.6
frozenlist==1.8.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
jiter==0.12.0
limits==3.8.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
nanoid==2.0.0
oauthlib==3.3.1
ollama==0.6.0
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-dbapi==0.59b0
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-flask==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-wsgi==0.59b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==23.2
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.2.1
pydantic_core==2.41.5
PyJWT==2.10.1
python-jose[cryptography]>=3.3.0
pyproject_hooks==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.18
PyYAML==6.0.1
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.36.3
tiktoken==0.8.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0.post1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
flask>=2.0.3
authlib>=1.0